microservices/webservice_for_Nurdin/requirements.txt
python-dateutil
Flask
uwsgi